Adding a fallback cpu_model to the node models config and adjusting IchnosCF/EmbodiedCarbon to use it
