+++ src/external_apis/Boavizta.py
@@ -12,6 +12,7 @@ def get_cpu_impact(cpu_name: str):
     Returns:
         float: The embedded GWP (Global Warming Potential) impact value of the CPU.
     """
+    print(f"Fetching CPU impact for {cpu_name}")
     url = "https://api.boavizta.org/v1/component/cpu?verbose=false"
     return make_json_post_request(url, {'name': cpu_name})['impacts']['gwp']['embedded']['value']
 
@@ -28,6 +29,7 @@ def get_aws_instance_impact(instance_type: str, duration: float = None):
     Returns:
         float: The embedded GWP (Global Warming Potential) impact value of the AWS instance.
     """
+    print(f"Fetching AWS instance impact for {instance_type} with duration {duration}")
     url = f"https://api.boavizta.org/v1/cloud/instance?provider=aws&instance_type={instance_type}&verbose=false&criteria=gwp"
     if duration: url += f"&duration={duration}"
     return make_json_get_request(url)['impacts']['gwp']['embedded']['value']
