+++ src/utils/PowerModel.py
@@ -3,7 +3,6 @@ import json
 import src.utils.MathModels as MathModels
 from src.Constants import DEFAULT_MEMORY_POWER_DRAW
 
-
 def get_power_model(model_name: str) -> Any:
     """
     Return a power model function requested based on stored configuration data
@@ -36,30 +35,3 @@ def get_power_model(model_name: str) -> Any:
             return MathModels.fitted_linear_power_model(coeff, inter)
 
         return MathModels.polynomial_model(models[node_id][governor][model_type])  # this should not be used...
-
-
-def get_memory_draw(model_name: str) -> Any:
-    try:
-        with open('node_config_models/nodes.json') as nodes_json_data:
-            models = json.load(nodes_json_data)
-
-            # Get the model data
-            model_data = model_name.split('_')
-            node_id: str = model_data[0]
-            governor: str = model_data[1]
-
-            return models[node_id][governor]['mem_draw']
-    except:
-        return DEFAULT_MEMORY_POWER_DRAW
-
-
-def get_system_cores(model_name: str) -> Any:
-    with open('node_config_models/nodes.json') as nodes_json_data:
-        models = json.load(nodes_json_data)
-
-        # Get the model data
-        model_data = model_name.split('_')
-        node_id: str = model_data[0]
-        governor: str = model_data[1]
-
-        return models[node_id][governor]['system_cores']
